requirements.txt
#required libraries
aiohttp==3.6.2
Pillow==7.1.1
python-multipart==0.0.5
starlette==0.13.3
uvicorn==0.11.3

keras==2.2.4
tensorflow==1.13.1